add StateChange and PushVerification entities
StateChange and PushVerification are two entities used in push.

Both have a '@type' attribute in serialized form that are pretty much useless
in a statically typed language. Therefor we don’t want to carry them around
as properties of our entities and instead only inject them during serialization.

The new TypeInjectAdapterFactory in combination with the @Type annotation takes
care of that.

Any entity annotated with `@Type` will have an additional `@type` key injected
into its JSON representation.
The value of the `@type` key can either be specified with `@Type("something")`
or else it will default to the class name.